+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log for SD.Next
 
+- fix startup issue with incorrect config
+- image viewer: fixes for non-chromium browsers
+- extra network: fix corrupt display on refesh when new extra network type found
+- diffusers: option to move base and/or refiner model to cpu to free up vram
+- diffusers: model downloader options to specify model variant/revision/mirror
+
 ## Update for 07/08/2023
 
 This is a massive update which has been baking in a `dev` branch for a while now